refactor: 사용하지 않는 Players, Player 클래스 제거
