+++ src/main/java/blackjack/domain/participant/Players2.java
@@ -13,7 +13,31 @@ public class Players2 {
     private final List<Player2> players;
 
     public Players2(List<Player2> players) {
+        validate(players);
         this.players = players;
+    }
+
+    private void validate(List<Player2> players) {
+        validateEntryNotEmpty(players);
+        validateEachPlayerNameUnique(players);
+    }
+
+    private void validateEntryNotEmpty(List<Player2> players) {
+        if (players.isEmpty()) {
+            throw new IllegalArgumentException("[ERROR] 플레이어가 없습니다");
+        }
+    }
+
+    private void validateEachPlayerNameUnique(List<Player2> players) {
+        if (countUniquePlayer(players) != players.size()) {
+            throw new IllegalArgumentException("[ERROR] 중복되는 플레이어의 이름이 존재합니다");
+        }
+    }
+
+    private int countUniquePlayer(List<Player2> players) {
+        return (int) players.stream()
+                .distinct()
+                .count();
     }
 
     public static Players2 create(List<String> rawNames) {
