+++ src/main/java/blackjack/view/MessageResolver.java
@@ -6,10 +6,7 @@ import blackjack.domain.game.PlayersResult;
 import blackjack.domain.game.Result2;
 import blackjack.domain.participant.Dealer2;
 import blackjack.domain.participant.Name;
-import blackjack.domain.participant.Player;
 import blackjack.domain.participant.Player2;
-import blackjack.domain.participant.PlayerName;
-import blackjack.domain.participant.Players;
 import blackjack.domain.participant.Players2;
 import blackjack.view.mapper.CardRankMapper;
 import blackjack.view.mapper.CardSuitMapper;
@@ -23,21 +20,9 @@ public class MessageResolver {
     private static final String DEALER_NAME = "딜러";
     private static final String HAND_FORMAT = "%s 카드: %s";
 
-    public String resolveDealDescriptionMessage(Players players) {
-        String message = String.format("%s와 %s에게 2장을 나누었습니다.", DEALER_NAME, resolveNamesMessage(players));
-        return String.join("", LINE_SEPARATOR, message);
-    }
-
     public String resolveDealDescriptionMessage(Players2 players) {
         String message = String.format("%s와 %s에게 2장을 나누었습니다.", DEALER_NAME, resolveNamesMessage(players));
         return String.join("", LINE_SEPARATOR, message);
-    }
-
-    private String resolveNamesMessage(Players players) {
-        return players.getPlayers().stream()
-                .map(Player::getPlayerName)
-                .map(PlayerName::getValue)
-                .collect(Collectors.joining(SEPARATOR));
     }
 
     private String resolveNamesMessage(Players2 players) {
