+++ src/test/java/blackjack/domain/participant/Players2Test.java
@@ -1,6 +1,8 @@
 package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
@@ -14,6 +16,33 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class Players2Test {
+
+    @DisplayName("플레이어가 없으면 예외가 발생한다.")
+    @Test
+    void testCreatePlayersWithEmptyEntry() {
+        assertThatThrownBy(() -> new Players2(List.of()))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("참가자들 중 이름이 중복되는 경우는 예외가 발생한다.")
+    @Test
+    void testCreatePlayersWithDuplicateNames() {
+        Player2 player1 = new Player2(new Name("pobi"));
+        Player2 player2 = new Player2(new Name("pobi"));
+
+        assertThatThrownBy(() -> new Players2(List.of(player1, player2)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("생성 검증을 모두 통과하면 생성에 성공한다.")
+    @Test
+    void testCreateWithValidPlayers() {
+        Player2 player1 = new Player2(new Name("pobi"));
+        Player2 player2 = new Player2(new Name("jason"));
+
+        assertThatCode(() -> new Players2(List.of(player1, player2)))
+                .doesNotThrowAnyException();
+    }
 
     @DisplayName("주어진 플레이어 이름 리스트로 플레이어들을 생성한다.")
     @Test
